Load customizations after Elpa runs
I was putting all the initialization of packages loaded by elpa and
el-get into my ~/.emacs.d/init.el file.  The problem is I can not
initialize the packages as though they have been loaded by placing the
code in there.  Done here use one of the many kluges around this issue
by putting the initialization in a file connected to the
after-init-hook.